requirements.txt
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
env==0.1.0
Flask==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
mohawk==1.1.0
python-dotenv==1.0.0
python-env==1.0.0
requests==2.31.0
six==1.16.0
SkPy==0.10.6
soupsieve==2.5
urllib3==2.1.0
Werkzeug==3.0.1